repackaged wrap_1d
