+++ autoarray/plot/wrap/two_d/mask_scatter.py
@@ -1,5 +1,6 @@
 from autoarray.plot.wrap.two_d.grid_scatter import GridScatter
 
+
 class MaskScatter(GridScatter):
     """
     Plots a mask over an image, using the `Mask2d` object's (y,x) `edge_grid_sub_1` property.
