+++ autoarray/plot/wrap/two_d/origin_scatter.py
@@ -1,5 +1,6 @@
 from autoarray.plot.wrap.two_d.grid_scatter import GridScatter
 
+
 class OriginScatter(GridScatter):
     """
     Plots the (y,x) coordinates of the origin of a data structure (e.g. as a black cross).
