+++ autoarray/plot/wrap/two_d/serial_overscan_plot.py
@@ -1,4 +1,5 @@
 from autoarray.plot.wrap.two_d.grid_plot import GridPlot
 
+
 class SerialOverscanPlot(GridPlot):
     pass
