+++ autoarray/structures/abstract_structure.py
